ETLs/postgres_to_es/requirements.txt
backoff==1.10.0
psycopg2-binary==2.8.6
pydantic==1.7.3
requests==2.25.1
urllib3==1.26.2
